added admin to nav bar
